requirements.txt
jupyter
matplotlib
numpy
pillow
scikit-image
scikit-learn
pandas

